ugly but progress
