--- apps/member-profile/app/routes/_profile.events._index.tsx
@@ -1,10 +1,0 @@
-import { type LoaderFunctionArgs, redirect } from '@remix-run/node';
-
-import { Route } from '@/shared/constants';
-import { ensureUserAuthenticated } from '@/shared/session.server';
-
-export async function loader({ request }: LoaderFunctionArgs) {
-  await ensureUserAuthenticated(request);
-
-  return redirect(Route['/events/upcoming']);
-}
