+++ apps/member-profile/app/shared/constants.ts
@@ -16,11 +16,9 @@ const ROUTES = [
   '/directory/join/finish',
   '/directory/:id',
   '/events',
-  '/events/past',
-  '/events/past/:id/attendees',
-  '/events/upcoming',
-  '/events/upcoming/:id/register',
-  '/events/upcoming/:id/registrations',
+  '/events/:id/attendees',
+  '/events/:id/register',
+  '/events/:id/registrations',
   '/home',
   '/home/activation',
   '/login',
